requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
cachetools==4.2.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
google-api-core==2.2.0
google-auth==2.3.2
google-cloud-vision==2.6.1
googleapis-common-protos==1.53.0
grpcio==1.41.1
grpcio-status==1.41.1
idna==3.3
multidict==5.2.0
packaging==21.0
proto-plus==1.19.7
protobuf==3.19.0
py-cord==2.0.0a4188+g9496c724
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.3
python-dotenv==0.19.1
python-Levenshtein==0.12.2
requests==2.26.0
rsa==4.7.2
six==1.16.0
thefuzz==0.19.0
typing-extensions==3.10.0.2
urllib3==1.26.7
yarl==1.7.0
